refactor(UIKIT-1275): Поправлены ошибки типов
