+++ packages/components/src/Profile/index.ts
@@ -1,3 +1,1 @@
 export * from './Profile';
-
-export * from './types';
