+++ packages/components/src/RadioGroup/RadioGroupContext/RadioGroupContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export type RadioGroupContextProps = {
+  isError?: boolean;
+};
+
+export const RadioGroupContext = createContext<RadioGroupContextProps>({
+  isError: false,
+});
